[GSF] Syncing Fork

+++ studio/src/app/handlers/editor/events/deck/deck-events.handler.tsx
@@ -18,6 +18,7 @@ import {DeckEditorService} from '../../../../services/editor/deck/deck-editor.se
 import {AuthService} from '../../../../services/auth/auth.service';
 import {DeckService} from '../../../../services/data/deck/deck.service';
 import {SlideService} from '../../../../services/data/slide/slide.service';
+import {ApiSlideService} from '../../../../services/api/slide/api.slide.service';
 
 export class DeckEventsHandler {
 
@@ -39,6 +40,8 @@ export class DeckEventsHandler {
     private deckService: DeckService;
     private slideService: SlideService;
 
+    private apiSlideService: ApiSlideService;
+
     constructor() {
         this.errorService = ErrorService.getInstance();
         this.busyService = BusyService.getInstance();
@@ -49,6 +52,8 @@ export class DeckEventsHandler {
 
         this.deckService = DeckService.getInstance();
         this.slideService = SlideService.getInstance();
+
+        this.apiSlideService = ApiSlideService.getInstance();
     }
 
     init(el: HTMLElement): Promise<void> {
@@ -457,15 +462,24 @@ export class DeckEventsHandler {
 
                 this.deckEditorService.watch().pipe(take(1)).subscribe(async (deck: Deck) => {
                     if (deck && deck.data) {
-                        await this.slideService.delete(deck.id, slideId);
+                        const slide: Slide = await this.slideService.get(deck.id, slideId);
 
-                        // TODO: Ultimately should be moved in a Cloud Function
-                        // Update list of slide in the deck
-                        if (deck.data.slides && deck.data.slides.indexOf(slideId) > -1) {
-                            deck.data.slides.splice(deck.data.slides.indexOf(slideId), 1);
+                        if (slide && slide.data) {
+                            // 1. If needed, delete slide in the API
+                            await this.deleteApiSlide(deck, slide);
 
-                            const updatedDeck: Deck = await this.deckService.update(deck);
-                            this.deckEditorService.next(updatedDeck);
+                            // TODO: no rush but ultimately maybe should we move step 2. and 3. below in a cloud function?
+
+                            // 2. Delete the slide in Firestore
+                            await this.slideService.delete(deck.id, slideId);
+
+                            // 3. Update list of slide in the deck (in Firestore)
+                            if (deck.data.slides && deck.data.slides.indexOf(slideId) > -1) {
+                                deck.data.slides.splice(deck.data.slides.indexOf(slideId), 1);
+
+                                const updatedDeck: Deck = await this.deckService.update(deck);
+                                this.deckEditorService.next(updatedDeck);
+                            }
                         }
                     }
 
@@ -479,6 +493,28 @@ export class DeckEventsHandler {
                 this.errorService.error(err);
                 this.busyService.deckBusy(false);
                 resolve();
+            }
+        });
+    }
+
+    private deleteApiSlide(deck: Deck, slide: Slide): Promise<void> {
+        return new Promise<void>(async (resolve, reject) => {
+            if (!deck.data.api_id || deck.data.api_id === undefined || deck.data.api_id === '') {
+                resolve();
+                return;
+            }
+
+            if (!slide.data.api_id || slide.data.api_id === undefined || slide.data.api_id === '') {
+                resolve();
+                return;
+            }
+
+            try {
+                await this.apiSlideService.delete(deck.data.api_id, slide.data.api_id);
+
+                resolve();
+            } catch (err) {
+                reject(err);
             }
         });
     }
